* Fix bug #43379,#43385.

+++ module/programplan/js/create.ui.js
@@ -14,7 +14,7 @@ window.onRenderRow = function(row, rowIdx, data)
             });
         }
 
-        row.find('[data-name="ACTIONS"]').find('[data-type="delete"]').remove();
+        if(data.hasOwnProperty('id')) row.find('[data-name="ACTIONS"]').find('[data-type="delete"]').remove();
     }
 
     if(!data || !data.planIDList) return;
